+++ src/ktcalendars/__init__.py
@@ -1,4 +1,5 @@
 from ktcalendars.calendar import KTCalendar
+from ktcalendars.config import AbstractConfiguration, DefaultConfiguration
 from ktcalendars.days import KTDay
 
-__all__ = ["KTCalendar", "KTDay"]
+__all__ = ["AbstractConfiguration", "DefaultConfiguration", "KTCalendar", "KTDay"]